Update .gitattributes
The increment function starts in the bottom right node and works its way to the bottom left, then it goes to the middle right, then works to the leftmost middle, then repeats again for the top row.  0 represents a null value and if there is a zero between two numbers the program will never end.